+++ app/src/main/kotlin/jp/cordea/mackerelclient/adapter/UserAdapter.kt
@@ -1,15 +1,14 @@
 package jp.cordea.mackerelclient.adapter
 
+import android.app.AlertDialog
 import android.content.Context
+import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.ImageView
 import android.widget.TextView
 import com.ogaclejapan.rx.binding.RxEvent
-import com.pawegio.kandroid.alert
-import com.pawegio.kandroid.find
-import com.pawegio.kandroid.inflateLayout
 import com.squareup.picasso.Picasso
 import jp.cordea.mackerelclient.PicassoCircularTransform
 import jp.cordea.mackerelclient.R
@@ -28,11 +27,11 @@ class UserAdapter(context: Context, val items: List<User>, val own: String?) : A
     public val onUserDeleteSucceed = RxEvent.create<Boolean>()
 
     override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View? {
-        val view = convertView ?: context.inflateLayout(R.layout.list_item_user, parent)
-        val name: TextView = view.find(R.id.name)
-        val email: TextView = view.find(R.id.email)
-        val delete: ImageView = view.find(R.id.delete_button)
-        val thumbnail: ImageView = view.find(R.id.user_thumbnail)
+        val view = convertView ?: LayoutInflater.from(context).inflate(R.layout.list_item_user, parent)
+        val name: TextView = view.findViewById(R.id.name) as TextView
+        val email: TextView = view.findViewById(R.id.email) as TextView
+        val delete: ImageView = view.findViewById(R.id.delete_button) as ImageView
+        val thumbnail: ImageView = view.findViewById(R.id.user_thumbnail) as ImageView
         GravatarUtils.getGravatarImage(items[position].email,
                 context.resources.getDimensionPixelSize(R.dimen.user_thumbnail_size_small))?.let { url ->
             Picasso.with(context)
@@ -51,40 +50,40 @@ class UserAdapter(context: Context, val items: List<User>, val own: String?) : A
         }
         if (!isOwn) {
             delete.setOnClickListener {
-                context.alert {
-                    message(R.string.user_delete_dialog_title)
-                    positiveButton(R.string.delete_positive_button, {
-                        val dialog = DialogUtils.progressDialog(context, R.string.progress_dialog_title)
-                        dialog.show()
-                        MackerelApiClient
-                                .deleteUser(context, items[position].id)
-                                .enqueue(object : Callback<User> {
-                                    override fun onResponse(response: Response<User>?) {
-                                        dialog.dismiss()
-                                        response?.let {
-                                            val success = DialogUtils.switchDialog(context, it,
-                                                    R.string.user_delete_error_dialog_title,
-                                                    R.string.error_403_dialog_message)
-                                            if (success) {
-                                                onUserDeleteSucceed.post(true)
+                AlertDialog
+                        .Builder(context)
+                        .setMessage(R.string.user_delete_dialog_title)
+                        .setPositiveButton(R.string.delete_positive_button, { dialogInterface, i ->
+                            val dialog = DialogUtils.progressDialog(context, R.string.progress_dialog_title)
+                            dialog.show()
+                            MackerelApiClient
+                                    .deleteUser(context, items[position].id)
+                                    .enqueue(object : Callback<User> {
+                                        override fun onResponse(response: Response<User>?) {
+                                            dialog.dismiss()
+                                            response?.let {
+                                                val success = DialogUtils.switchDialog(context, it,
+                                                        R.string.user_delete_error_dialog_title,
+                                                        R.string.error_403_dialog_message)
+                                                if (success) {
+                                                    onUserDeleteSucceed.post(true)
+                                                }
+                                                return
                                             }
-                                            return
+                                            DialogUtils.showDialog(context,
+                                                    R.string.user_delete_error_dialog_title)
                                         }
-                                        DialogUtils.showDialog(context,
-                                                R.string.user_delete_error_dialog_title)
-                                    }
 
-                                    override fun onFailure(t: Throwable?) {
-                                        dialog.dismiss()
-                                        DialogUtils.showDialog(context,
-                                                R.string.user_delete_error_dialog_title)
-                                    }
-                                })
-                    })
-                }.show()
+                                        override fun onFailure(t: Throwable?) {
+                                            dialog.dismiss()
+                                            DialogUtils.showDialog(context,
+                                                    R.string.user_delete_error_dialog_title)
+                                        }
+                                    })
+                        })
+                .show()
             }
         }
-
         return view
     }
